% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paneledOutlierExplorer.R
\name{paneledOutlierExplorer}
\alias{paneledOutlierExplorer}
\title{Paneled Outlier Explorer}
\usage{
paneledOutlierExplorer(data, width = NULL, height = NULL, elementId = NULL)
}
\description{
Paneled Outlier Explorer
}
